// mobile/lib/types.ts
export interface User {
    id: string;
    name: string;
    email: string;
    avatarUrl?: string;
    bio?: string;
    createdAt: string;
    xp: number;
    level: number;
    loginStreak: number;
}
// ...
export interface FriendUser extends User {
    friendStatus: 'none' | 'pending' | 'accepted' | 'rejected';
    friendshipId?: string | null;
}

export interface FriendRequest {
    id: string;
    requesterId: string;
    addresseeId: string;
    status: 'pending' | 'accepted' | 'rejected';
    createdAt: string;
    updatedAt?: string;
    requester: User;
}